Split common into field related files
